requirements.txt
git+https://github.com/facebookresearch/pytorch3d.git
matplotlib
numpy
pyrender
trimesh
cvxpy
scipy
imageio
plyfile
scikit-image
pytest
pyhocon
pyyaml
pillow
sacred
cython<3
setuptools
py3ode
PyYAML
icecream
opencv-python
cprint
open3d
zarr
klampt
pyrender
pygame
pykeops
pyhocon
libigl
rtree
plyfile==0.7
scikit-image>=0.18.0
ipdb==0.13.7
torch-scatter 
-f https://data.pyg.org/whl/torch-2.1.0+cu121.html